backend/pyproject.toml
[tool.poetry]
name = "todo-app-practice-1-backend"
description = ""
authors = ["Mao Matsumoto"]
version = "0.1.0"

[tool.poetry.dependencies]
boto3 = "1.34.44"
fastapi = "0.109.2"
mangum = "0.17.0"
mypy-boto3-dynamodb = "1.34.46"
pydantic = {extras = ["dotenv"], version = "2.6.1"}
pydantic-settings = "2.2.0"
python = "^3.12"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.group.dev.dependencies]
black = "24.2.0"
httpx = "0.27.0"
poethepoet = "0.24.4"
pytest = "8.0.0"
pytest-asyncio = "0.23.5"
pytest-cov = "4.1.0"
ruff = "0.2.1"
uvicorn = {extras = ["standard"], version = "0.27.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.black]
line-length = 120

[tool.coverage.run]
branch = true
source = ["src"]
data_file = ".pytest_cache/.coverage"

[tool.coverage.report]
exclude_lines = []

[tool.coverage.xml]
output = ".pytest_cache/coverage.xml"

[tool.poe.tasks]
fmt = "black src tests --check"
lint = "ruff check src tests"
test = "pytest --junitxml=./.pytest_cache/pytest.xml --cov --cov-report xml"
dev = "uvicorn src.main:app --host 0.0.0.0 --port 8000 --reload"
